start work on better type system for dependent types
